@@ -43,6 +43,7 @@ class SwiftFlexibleSyncTests: SwiftSyncTestCase {
         try createFlexibleSyncApp()
     }
 
+    @MainActor
     func testCreateFlexibleSyncApp() throws {
         let appId = try RealmServer.shared.createApp(fields: ["age"], types: [SwiftPerson.self])
         let flexibleApp = app(id: appId)
@@ -62,18 +63,21 @@ class SwiftFlexibleSyncTests: SwiftSyncTestCase {
         assertThrows(realm.subscriptions)
     }
 
+    @MainActor
     func testFlexibleSyncPath() throws {
         let config = try configuration()
         let user = config.syncConfiguration!.user
         XCTAssertTrue(config.fileURL!.path.hasSuffix("mongodb-realm/\(appId)/\(user.id)/flx_sync_default.realm"))
     }
 
+    @MainActor
     func testGetSubscriptions() throws {
         let realm = try openRealm()
         let subscriptions = realm.subscriptions
         XCTAssertEqual(subscriptions.count, 0)
     }
 
+    @MainActor
     func testWriteEmptyBlock() throws {
         let realm = try openRealm()
         let subscriptions = realm.subscriptions
@@ -82,6 +86,7 @@ class SwiftFlexibleSyncTests: SwiftSyncTestCase {
         XCTAssertEqual(subscriptions.count, 0)
     }
 
+    @MainActor
     func testAddOneSubscriptionWithoutName() throws {
         let realm = try openRealm()
         let subscriptions = realm.subscriptions
@@ -94,6 +99,7 @@ class SwiftFlexibleSyncTests: SwiftSyncTestCase {
         XCTAssertEqual(subscriptions.count, 1)
     }
 
+    @MainActor
     func testAddOneSubscriptionWithName() throws {
         let realm = try openRealm()
         let subscriptions = realm.subscriptions
@@ -106,6 +112,7 @@ class SwiftFlexibleSyncTests: SwiftSyncTestCase {
         XCTAssertEqual(subscriptions.count, 1)
     }
 
+    @MainActor
     func testAddSubscriptionsInDifferentBlocks() throws {
         let realm = try openRealm()
         let subscriptions = realm.subscriptions
@@ -123,6 +130,7 @@ class SwiftFlexibleSyncTests: SwiftSyncTestCase {
         XCTAssertEqual(subscriptions.count, 2)
     }
 
+    @MainActor
     func testAddSeveralSubscriptionsWithoutName() throws {
         let realm = try openRealm()
         let subscriptions = realm.subscriptions
@@ -142,6 +150,7 @@ class SwiftFlexibleSyncTests: SwiftSyncTestCase {
         XCTAssertEqual(subscriptions.count, 3)
     }
 
+    @MainActor
     func testAddSeveralSubscriptionsWithName() throws {
         let realm = try openRealm()
         let subscriptions = realm.subscriptions
@@ -160,6 +169,7 @@ class SwiftFlexibleSyncTests: SwiftSyncTestCase {
         XCTAssertEqual(subscriptions.count, 3)
     }
 
+    @MainActor
     func testAddMixedSubscriptions() throws {
         let realm = try openRealm()
         let subscriptions = realm.subscriptions
@@ -178,6 +188,7 @@ class SwiftFlexibleSyncTests: SwiftSyncTestCase {
         XCTAssertEqual(subscriptions.count, 3)
     }
 
+    @MainActor
     func testAddDuplicateSubscriptions() throws {
         let realm = try openRealm()
         let subscriptions = realm.subscriptions
@@ -193,6 +204,7 @@ class SwiftFlexibleSyncTests: SwiftSyncTestCase {
         XCTAssertEqual(subscriptions.count, 1)
     }
 
+    @MainActor
     func testAddDuplicateSubscriptionWithDifferentName() throws {
         let realm = try openRealm()
         let subscriptions = realm.subscriptions
@@ -209,6 +221,7 @@ class SwiftFlexibleSyncTests: SwiftSyncTestCase {
     }
 
     // FIXME: Using `assertThrows` within a Server test will crash on tear down
+    @MainActor
     func skip_testSameNamedSubscriptionThrows() throws {
         let realm = try openRealm()
         let subscriptions = realm.subscriptions
@@ -224,6 +237,7 @@ class SwiftFlexibleSyncTests: SwiftSyncTestCase {
     }
 
     // FIXME: Using `assertThrows` within a Server test will crash on tear down
+    @MainActor
     func skip_testAddSubscriptionOutsideWriteThrows() throws {
         let realm = try openRealm()
         let subscriptions = realm.subscriptions
@@ -232,6 +246,7 @@ class SwiftFlexibleSyncTests: SwiftSyncTestCase {
         }))
     }
 
+    @MainActor
     func testFindSubscriptionByName() throws {
         let realm = try openRealm()
         let subscriptions = realm.subscriptions
@@ -255,6 +270,7 @@ class SwiftFlexibleSyncTests: SwiftSyncTestCase {
         XCTAssertEqual(foundSubscription2!.name, "person_age_20")
     }
 
+    @MainActor
     func testFindSubscriptionByQuery() throws {
         let realm = try openRealm()
         let subscriptions = realm.subscriptions
@@ -281,6 +297,7 @@ class SwiftFlexibleSyncTests: SwiftSyncTestCase {
         XCTAssertEqual(foundSubscription2!.name, "object_int_more_than_zero")
     }
 
+    @MainActor
     func testRemoveSubscriptionByName() throws {
         let realm = try openRealm()
         let subscriptions = realm.subscriptions
@@ -304,6 +321,7 @@ class SwiftFlexibleSyncTests: SwiftSyncTestCase {
         XCTAssertEqual(subscriptions.count, 2)
     }
 
+    @MainActor
     func testRemoveSubscriptionByQuery() throws {
         let realm = try openRealm()
         let subscriptions = realm.subscriptions
@@ -338,6 +356,7 @@ class SwiftFlexibleSyncTests: SwiftSyncTestCase {
         XCTAssertEqual(subscriptions.count, 1)
     }
 
+    @MainActor
     func testRemoveSubscription() throws {
         let realm = try openRealm()
         let subscriptions = realm.subscriptions
@@ -370,6 +389,7 @@ class SwiftFlexibleSyncTests: SwiftSyncTestCase {
         XCTAssertEqual(subscriptions.count, 0)
     }
 
+    @MainActor
     func testRemoveSubscriptionsByType() throws {
         let realm = try openRealm()
         let subscriptions = realm.subscriptions
@@ -401,6 +421,7 @@ class SwiftFlexibleSyncTests: SwiftSyncTestCase {
         XCTAssertEqual(subscriptions.count, 0)
     }
 
+    @MainActor
     func testRemoveAllSubscriptions() throws {
         let realm = try openRealm()
         let subscriptions = realm.subscriptions
@@ -428,6 +449,7 @@ class SwiftFlexibleSyncTests: SwiftSyncTestCase {
         XCTAssertEqual(subscriptions.count, 0)
     }
 
+    @MainActor
     func testRemoveAllUnnamedSubscriptions() throws {
         let realm = try openRealm()
         let subscriptions = realm.subscriptions
@@ -455,6 +477,7 @@ class SwiftFlexibleSyncTests: SwiftSyncTestCase {
         XCTAssertEqual(subscriptions.count, 2)
     }
 
+    @MainActor
     func testSubscriptionSetIterate() throws {
         let realm = try openRealm()
         let subscriptions = realm.subscriptions
@@ -479,6 +502,7 @@ class SwiftFlexibleSyncTests: SwiftSyncTestCase {
         XCTAssertEqual(count, numberOfSubs)
     }
 
+    @MainActor
     func testSubscriptionSetFirstAndLast() throws {
         let realm = try openRealm()
         let subscriptions = realm.subscriptions
@@ -503,6 +527,7 @@ class SwiftFlexibleSyncTests: SwiftSyncTestCase {
         XCTAssertEqual(lastSubscription!.name, "person_age_\(numberOfSubs)")
     }
 
+    @MainActor
     func testSubscriptionSetSubscript() throws {
         let realm = try openRealm()
         let subscriptions = realm.subscriptions
@@ -527,6 +552,7 @@ class SwiftFlexibleSyncTests: SwiftSyncTestCase {
         XCTAssertEqual(lastSubscription!.name, "person_age_\(numberOfSubs)")
     }
 
+    @MainActor
     func testUpdateQueries() throws {
         let realm = try openRealm()
         let subscriptions = realm.subscriptions
@@ -552,6 +578,7 @@ class SwiftFlexibleSyncTests: SwiftSyncTestCase {
         XCTAssertEqual(subscriptions.count, 2)
     }
 
+    @MainActor
     func testUpdateQueriesWithoutName() throws {
         let realm = try openRealm()
         let subscriptions = realm.subscriptions
@@ -578,6 +605,7 @@ class SwiftFlexibleSyncTests: SwiftSyncTestCase {
     }
 
     // FIXME: Using `assertThrows` within a Server test will crash on tear down
+    @MainActor
     func skip_testFlexibleSyncAppUpdateQueryWithDifferentObjectTypeWillThrow() throws {
         let realm = try openRealm()
         let subscriptions = realm.subscriptions
@@ -596,6 +624,7 @@ class SwiftFlexibleSyncTests: SwiftSyncTestCase {
         }
     }
 
+    @MainActor
     func testFlexibleSyncTransactionsWithPredicateFormatAndNSPredicate() throws {
         let realm = try openRealm()
         let subscriptions = realm.subscriptions
@@ -624,6 +653,7 @@ class SwiftFlexibleSyncTests: SwiftSyncTestCase {
         XCTAssertEqual(subscriptions.count, 2)
     }
 
+    @MainActor
     func populateSwiftPerson(count: Int = 10) throws {
         try write { realm in
             for i in 1...count {
@@ -632,6 +662,7 @@ class SwiftFlexibleSyncTests: SwiftSyncTestCase {
         }
     }
 
+    @MainActor
     func populateSwiftTypesObject(count: Int = 1) throws {
         try write { realm in
             for _ in 1...count {
@@ -642,6 +673,7 @@ class SwiftFlexibleSyncTests: SwiftSyncTestCase {
         }
     }
 
+    @MainActor
     func testFlexibleSyncAppWithoutQuery() throws {
         try populateSwiftPerson()
 
@@ -655,6 +687,7 @@ class SwiftFlexibleSyncTests: SwiftSyncTestCase {
         checkCount(expected: 0, realm, SwiftPerson.self)
     }
 
+    @MainActor
     func testFlexibleSyncAppAddQuery() throws {
         try populateSwiftPerson(count: 25)
 
@@ -680,6 +713,7 @@ class SwiftFlexibleSyncTests: SwiftSyncTestCase {
         checkCount(expected: 10, realm, SwiftPerson.self)
     }
 
+    @MainActor
     func testFlexibleSyncAppMultipleQuery() throws {
         try populateSwiftPerson(count: 20)
         try populateSwiftTypesObject()
@@ -709,6 +743,7 @@ class SwiftFlexibleSyncTests: SwiftSyncTestCase {
         checkCount(expected: 1, realm, SwiftTypesSyncObject.self)
     }
 
+    @MainActor
     func testFlexibleSyncAppRemoveQuery() throws {
         try populateSwiftPerson(count: 30)
         try populateSwiftTypesObject()
@@ -751,6 +786,7 @@ class SwiftFlexibleSyncTests: SwiftSyncTestCase {
         checkCount(expected: 1, realm, SwiftTypesSyncObject.self)
     }
 
+    @MainActor
     func testFlexibleSyncAppRemoveAllQueries() throws {
         try populateSwiftPerson(count: 25)
         try populateSwiftTypesObject()
@@ -797,6 +833,7 @@ class SwiftFlexibleSyncTests: SwiftSyncTestCase {
         checkCount(expected: 0, realm, SwiftTypesSyncObject.self)
     }
 
+    @MainActor
     func testFlexibleSyncAppRemoveQueriesByType() throws {
         try populateSwiftPerson(count: 21)
         try populateSwiftTypesObject()
@@ -843,6 +880,7 @@ class SwiftFlexibleSyncTests: SwiftSyncTestCase {
         checkCount(expected: 1, realm, SwiftTypesSyncObject.self)
     }
 
+    @MainActor
     func testFlexibleSyncAppUpdateQuery() throws {
         try populateSwiftPerson(count: 25)
 
@@ -884,13 +922,15 @@ class SwiftFlexibleSyncTests: SwiftSyncTestCase {
         checkCount(expected: 20, realm, SwiftPerson.self)
     }
 
+    @MainActor
     func testFlexibleSyncInitialSubscriptions() throws {
         try populateSwiftPerson(count: 20)
 
+        let name = self.name
         let user = createUser()
         var config = user.flexibleSyncConfiguration(initialSubscriptions: { subscriptions in
             subscriptions.append(QuerySubscription<SwiftPerson>(name: "person_age_10") {
-                $0.age > 10 && $0.firstName == self.name
+                $0.age > 10 && $0.firstName == name
             })
         })
         config.objectTypes = [SwiftPerson.self, SwiftTypesSyncObject.self]
@@ -910,6 +950,7 @@ class SwiftFlexibleSyncTests: SwiftSyncTestCase {
         checkCount(expected: 10, realm, SwiftPerson.self)
     }
 
+    @MainActor
     func testFlexibleSyncCancelOnNonFatalError() throws {
         let proxy = TimeoutProxyServer(port: 5678, targetPort: 9090)
         try proxy.start()
@@ -943,15 +984,17 @@ class SwiftFlexibleSyncTests: SwiftSyncTestCase {
     }
 
     // MARK: - Progress notifiers
+    @MainActor
     func testAsyncOpenProgress() throws {
         try populateRealm()
 
         let asyncOpenEx = expectation(description: "async open")
 
         let user = createUser()
+        let name = self.name
         var config = user.flexibleSyncConfiguration(initialSubscriptions: { subscriptions in
             subscriptions.append(QuerySubscription<SwiftHugeSyncObject> {
-                $0.partition == self.name
+                $0.partition == name
             })
         })
         config.objectTypes = objectTypes
@@ -984,6 +1027,7 @@ class SwiftFlexibleSyncTests: SwiftSyncTestCase {
         XCTAssertTrue(p1.isTransferComplete)
     }
 
+    @MainActor
     func testNonStreamingDownloadNotifier() async throws {
         try populateRealm()
 
@@ -1036,6 +1080,7 @@ class SwiftFlexibleSyncTests: SwiftSyncTestCase {
         token!.invalidate()
     }
 
+    @MainActor
     func testStreamingDownloadNotifier() throws {
         try populateRealm()
 
@@ -1093,6 +1138,7 @@ class SwiftFlexibleSyncTests: SwiftSyncTestCase {
         token!.invalidate()
     }
 
+    @MainActor
     func testStreamingUploadNotifier() throws {
         let realm = try openRealm(wait: false)
         let subscriptions = realm.subscriptions
@@ -1136,6 +1182,7 @@ class SwiftFlexibleSyncTests: SwiftSyncTestCase {
         XCTAssertTrue(p.isTransferComplete)
     }
 
+    @MainActor
     func testStreamingNotifierInvalidate() throws {
         let realm = try openRealm()
         RLMRealmSubscribeToAll(ObjectiveCSupport.convert(object: realm))
@@ -1187,6 +1234,7 @@ class SwiftFlexibleSyncTests: SwiftSyncTestCase {
         XCTAssertEqual(uploadCount.value, 0)
     }
 
+    @MainActor
     func testFlexibleSyncNotEnabledError() throws {
         let appId = try RealmServer.shared.createNonSyncApp()
         let app = app(id: appId)
