@@ -64,7 +64,7 @@ public class TimeoutProxyServer: NSObject, @unchecked Sendable {
 
     @objc public func start() throws {
         listener = try NWListener(using: NWParameters.tcp, on: port)
-        listener.newConnectionHandler = { [weak self] incomingConnection in
+        listener.newConnectionHandler = { @Sendable [weak self] incomingConnection in
             guard let self = self else { return }
             self.connections.append(incomingConnection)
             incomingConnection.start(queue: self.queue)
