@@ -150,10 +150,12 @@ class OnlyComputedNoBacklinksProps: FakeObject {
     }
 }
 
-@MainActor
 class RequiresObjcName: RLMObject {
+#if compiler(>=5.10)
+    nonisolated(unsafe) static var enable = false
+#else
     static var enable = false
-    @MainActor
+#endif
     override class func _realmIgnoreClass() -> Bool {
         return !enable
     }
