@@ -520,6 +520,7 @@ public final class Map<Key: _MapKey, Value: RealmCollectionValue>: RLMSwiftColle
         return rlmDictionary.addNotificationBlock(wrapped, keyPaths: keyPaths, queue: queue)
     }
 
+#if compiler(<6)
     /**
     Registers a block to be called each time the map changes.
 
@@ -609,7 +610,7 @@ public final class Map<Key: _MapKey, Value: RealmCollectionValue>: RLMSwiftColle
             collection.observe(keyPaths: keyPaths, on: nil) { change in
                 actor.invokeIsolated(block, change)
             }
-        } ?? NotificationToken()
+        }
     }
 
     /**
@@ -696,6 +697,185 @@ public final class Map<Key: _MapKey, Value: RealmCollectionValue>: RLMSwiftColle
     ) async -> NotificationToken where Value: OptionalProtocol, Value.Wrapped: ObjectBase {
         await observe(keyPaths: keyPaths.map(_name(for:)), on: actor, block)
     }
+#else
+    /**
+    Registers a block to be called each time the map changes.
+
+    The block will be asynchronously called on the actor with the initial map, and
+    then called again after each write transaction which changes either which keys
+    are present in the map or the values of any of the objects.
+
+    The `change` parameter that is passed to the block reports, in the form of keys
+    within the map, which of the key-value pairs were added, removed, or modified
+    during each write transaction.
+
+    Notifications are delivered to a function isolated to the given actor, on that
+    actors executor. If the actor is performing blocking work, multiple
+    notifications may be coalesced into a single notification. This can include the
+    notification with the initial collection, and changes are only reported for
+    writes which occur after the initial notification is delivered.
+
+    If no key paths are given, the block will be executed on any insertion,
+    modification, or deletion for all object properties and the properties of any
+    nested, linked objects. If a key path or key paths are provided, then the block
+    will be called for changes which occur only on the provided key paths. For
+    example, if:
+    ```swift
+    class Dog: Object {
+        @Persisted var name: String
+        @Persisted var age: Int
+        @Persisted var toys: List<Toy>
+    }
+    // ...
+    let dogs = myObject.mapOfDogs
+    let token = dogs.observe(keyPaths: ["name"], on: actor) { actor, changes in
+        switch changes {
+        case .initial(let dogs):
+            // ...
+        case .update:
+            // This case is hit:
+            // - after the token is initialized
+            // - when the name property of an object in the collection is modified
+            // - when an element is inserted or removed from the collection.
+            // This block is not triggered:
+            // - when a value other than name is modified on one of the elements.
+        case .error:
+            // No longer possible and left for backwards compatibility
+        }
+    }
+    ```
+    - If the observed key path were `["toys.brand"]`, then any insertion or
+      deletion to the `toys` list on any of the collection's elements would trigger
+      the block. Changes to the `brand` value on any `Toy` that is linked to a `Dog`
+      in this collection will trigger the block. Changes to a value other than
+      `brand` on any `Toy` that is linked to a `Dog` in this collection would not
+      trigger the block. Any insertion or removal to the `Dog` type collection being
+      observed would also trigger a notification.
+    - If the above example observed the `["toys"]` key path, then any insertion,
+      deletion, or modification to the `toys` list for any element in the collection
+      would trigger the block. Changes to any value on any `Toy` that is linked to a
+      `Dog` in this collection would *not* trigger the block. Any insertion or
+      removal to the `Dog` type collection being observed would still trigger a
+      notification.
+
+    You must retain the returned token for as long as you want updates to be sent
+    to the block. To stop receiving updates, call `invalidate()` on the token.
+
+    - warning: This method cannot be called during a write transaction, or when
+      the containing Realm is read-only.
+
+    - parameter keyPaths: Only properties contained in the key paths array will
+      trigger the block when they are modified. If `nil`, notifications will be
+      delivered for any property change on the object. String key paths which do not
+      correspond to a valid a property will throw an exception. See description above
+      for more detail on linked properties.
+    - note: The keyPaths parameter refers to object properties of the collection
+      type and *does not* refer to particular key/value pairs within the Map.
+    - parameter actor: The actor which notifications should be delivered on. The
+      block is passed this actor as an isolated parameter, allowing you to access the
+      actor synchronously from within the callback.
+    - parameter block: The block to be called whenever a change occurs.
+    - returns: A token which must be held for as long as you want updates to be delivered.
+     */
+    @available(macOS 10.15, tvOS 13.0, iOS 13.0, watchOS 6.0, *)
+    public func observe<A: Actor>(
+        keyPaths: [String]? = nil, on actor: A,
+        _isolation: isolated (any Actor)? = #isolation,
+        _ block: @Sendable @escaping (isolated A, RealmMapChange<Map>) -> Void
+    ) async -> NotificationToken {
+        nonisolated(unsafe) let collection = self
+        return await with(collection, on: actor) { actor, collection in
+            collection.observe(keyPaths: keyPaths, on: nil) { change in
+                actor.invokeIsolated(block, change)
+            }
+        }
+    }
+
+    /**
+    Registers a block to be called each time the map changes.
+
+    The block will be asynchronously called on the actor with the initial map, and
+    then called again after each write transaction which changes either which keys
+    are present in the map or the values of any of the objects.
+
+    The `change` parameter that is passed to the block reports, in the form of keys
+    within the map, which of the key-value pairs were added, removed, or modified
+    during each write transaction.
+
+    Notifications are delivered to a function isolated to the given actor, on that
+    actors executor. If the actor is performing blocking work, multiple
+    notifications may be coalesced into a single notification. This can include the
+    notification with the initial collection, and changes are only reported for
+    writes which occur after the initial notification is delivered.
+
+    The block will be called for changes which occur only on the provided key
+    paths. For example, if:
+    ```swift
+    class Dog: Object {
+        @Persisted var name: String
+        @Persisted var age: Int
+        @Persisted var toys: List<Toy>
+    }
+    // ...
+    let dogs = myObject.mapOfDogs
+    let token = dogs.observe(keyPaths: [\.name], on: actor) { actor, changes in
+        switch changes {
+        case .initial(let dogs):
+            // ...
+        case .update:
+            // This case is hit:
+            // - after the token is initialized
+            // - when the name property of an object in the collection is modified
+            // - when an element is inserted or removed from the collection.
+            // This block is not triggered:
+            // - when a value other than name is modified on one of the elements.
+        case .error:
+            // No longer possible and left for backwards compatibility
+        }
+    }
+    ```
+    - If the observed key path were `[\.toys.brand]`, then any insertion or
+      deletion to the `toys` list on any of the collection's elements would trigger
+      the block. Changes to the `brand` value on any `Toy` that is linked to a `Dog`
+      in this collection will trigger the block. Changes to a value other than
+      `brand` on any `Toy` that is linked to a `Dog` in this collection would not
+      trigger the block. Any insertion or removal to the `Dog` type collection being
+      observed would also trigger a notification.
+    - If the above example observed the `[\.toys]` key path, then any insertion,
+      deletion, or modification to the `toys` list for any element in the collection
+      would trigger the block. Changes to any value on any `Toy` that is linked to a
+      `Dog` in this collection would *not* trigger the block. Any insertion or
+      removal to the `Dog` type collection being observed would still trigger a
+      notification.
+
+    You must retain the returned token for as long as you want updates to be sent
+    to the block. To stop receiving updates, call `invalidate()` on the token.
+
+    - warning: This method cannot be called during a write transaction, or when
+      the containing Realm is read-only.
+
+    - parameter keyPaths: Only properties contained in the key paths array will
+      trigger the block when they are modified. If `nil`, notifications will be
+      delivered for any property change on the object. String key paths which do not
+      correspond to a valid a property will throw an exception. See description above
+      for more detail on linked properties.
+    - note: The keyPaths parameter refers to object properties of the collection
+      type and *does not* refer to particular key/value pairs within the Map.
+    - parameter actor: The actor which notifications should be delivered on. The
+      block is passed this actor as an isolated parameter, allowing you to access the
+      actor synchronously from within the callback.
+    - parameter block: The block to be called whenever a change occurs.
+    - returns: A token which must be held for as long as you want updates to be delivered.
+     */
+    @available(macOS 10.15, tvOS 13.0, iOS 13.0, watchOS 6.0, *)
+    public func observe<A: Actor>(
+        keyPaths: [PartialKeyPath<Value.Wrapped>], on actor: A,
+        _isolation: isolated (any Actor)? = #isolation,
+        _ block: @Sendable @escaping (isolated A, RealmMapChange<Map>) -> Void
+    ) async -> NotificationToken where Value: OptionalProtocol, Value.Wrapped: ObjectBase {
+        await observe(keyPaths: keyPaths.map(_name(for:)), on: actor, block)
+    }
+#endif
 
     // MARK: Frozen Objects
 
