@@ -603,12 +603,7 @@ extension MongoCollection {
             .compactMap(ObjectiveCSupport.convertBson(object:))
     }
 
-    // These uses of `@_unsafeInheritExecutor` should instead be marking the
-    // options parameters as `@Copy`. Unfortunately, as of Swift 5.8 that doesn't
-    // actually work due to https://github.com/apple/swift/issues/61358
-
-    // NEXT-MAJOR: make the options parameter non-optional and default to .init()
-    // instead of nil with nil-coalescing internally.
+#if compiler(<6)
     /// Finds the documents in this collection which match the provided filter.
     /// - Parameters:
     ///   - filter: A `Document` as bson that should match the query.
@@ -635,6 +630,48 @@ extension MongoCollection {
                                          options: options ?? .init())
             .map(ObjectiveCSupport.convert)
     }
+#else
+    /// Finds the documents in this collection which match the provided filter.
+    /// - Parameters:
+    ///   - filter: A `Document` as bson that should match the query.
+    ///   - options: `FindOptions` to use when executing the command.
+    /// - Returns: Array of `Document` filtered.
+    public func find(filter: Document, options: FindOptions = .init(),
+                     _isolation: isolated (any Actor)? = #isolation) async throws -> [Document] {
+        try await withCheckedThrowingContinuation { continuation in
+            __findWhere(ObjectiveCSupport.convert(filter),
+                        options: options) { bson, error in
+                if let error {
+                    continuation.resume(throwing: error)
+                } else {
+                    continuation.resume(returning: bson!.map(ObjectiveCSupport.convert))
+                }
+            }
+        }
+    }
+
+    /// Returns one document from a collection or view which matches the
+    /// provided filter. If multiple documents satisfy the query, this method
+    /// returns the first document according to the query's sort order or natural
+    /// order.
+    /// - Parameters:
+    ///   - filter: A `Document` as bson that should match the query.
+    ///   - options: `FindOptions` to use when executing the command.
+    /// - Returns: `Document` filtered.
+    public func findOneDocument(filter: Document, options: FindOptions = .init(),
+                                _isolation: isolated (any Actor)? = #isolation) async throws -> Document? {
+        try await withCheckedThrowingContinuation { continuation in
+            __findOneDocumentWhere(ObjectiveCSupport.convert(filter),
+                                   options: options) { bson, error in
+                if let error {
+                    continuation.resume(throwing: error)
+                } else {
+                    continuation.resume(returning: bson.map(ObjectiveCSupport.convert))
+                }
+            }
+        }
+    }
+#endif
 
     /// Runs an aggregation framework pipeline against this collection.
     /// - Parameters:
@@ -699,8 +736,7 @@ extension MongoCollection {
                                              upsert: upsert ?? false)
     }
 
-    // NEXT-MAJOR: make the options parameter non-optional and default to .init()
-    // instead of nil with nil-coalescing internally.
+#if compiler(<6)
     /// Updates a single document in a collection based on a query filter and
     /// returns the document in either its pre-update or post-update form. Unlike
     /// `updateOneDocument`, this action allows you to atomically find, update, and
@@ -758,6 +794,85 @@ extension MongoCollection {
                                           options: options ?? .init())
             .map(ObjectiveCSupport.convert)
     }
+#else
+    /// Updates a single document in a collection based on a query filter and
+    /// returns the document in either its pre-update or post-update form. Unlike
+    /// `updateOneDocument`, this action allows you to atomically find, update, and
+    /// return a document with the same command. This avoids the risk of other
+    /// update operations changing the document between separate find and update
+    /// operations.
+    /// - Parameters:
+    ///   - filter: A bson `Document` representing the match criteria.
+    ///   - update: A bson `Document` representing the update to be applied to a matching document.
+    ///   - options: `RemoteFindOneAndModifyOptions` to use when executing the command.
+    /// - Returns: `Document` result of the attempt to update a document  or `nil` if document wasn't found.
+    public func findOneAndUpdate(filter: Document, update: Document,
+                                 options: FindOneAndModifyOptions = .init(),
+                                 _isolation: isolated (any Actor)? = #isolation) async throws -> Document? {
+        try await withCheckedThrowingContinuation { continuation in
+            __findOneAndUpdateWhere(ObjectiveCSupport.convert(filter),
+                                    updateDocument: ObjectiveCSupport.convert(update),
+                                    options: options) { bson, error in
+                if let error {
+                    continuation.resume(throwing: error)
+                } else {
+                    continuation.resume(returning: bson.map(ObjectiveCSupport.convert))
+                }
+            }
+        }
+    }
+
+    /// Overwrites a single document in a collection based on a query filter and
+    /// returns the document in either its pre-replacement or post-replacement
+    /// form. Unlike `updateOneDocument`, this action allows you to atomically find,
+    /// replace, and return a document with the same command. This avoids the
+    /// risk of other update operations changing the document between separate
+    /// find and update operations.
+    /// - Parameters:
+    ///   - filter: A `Document` that should match the query.
+    ///   - replacement: A `Document` describing the replacement.
+    ///   - options: `FindOneAndModifyOptions` to use when executing the command.
+    /// - Returns: `Document`result of the attempt to reaplce a document   or `nil` if document wasn't found.
+    public func findOneAndReplace(filter: Document, replacement: Document,
+                                  options: FindOneAndModifyOptions = .init(),
+                                  _isolation: isolated (any Actor)? = #isolation) async throws -> Document? {
+        try await withCheckedThrowingContinuation { continuation in
+            __findOneAndReplaceWhere(ObjectiveCSupport.convert(filter),
+                                     replacementDocument: ObjectiveCSupport.convert(replacement),
+                                     options: options) { bson, error in
+                if let error {
+                    continuation.resume(throwing: error)
+                } else {
+                    continuation.resume(returning: bson.map(ObjectiveCSupport.convert))
+                }
+            }
+        }
+    }
+    /// Removes a single document from a collection based on a query filter and
+    /// returns a document with the same form as the document immediately before
+    /// it was deleted. Unlike `deleteOneDocument`, this action allows you to atomically
+    /// find and delete a document with the same command. This avoids the risk of
+    /// other update operations changing the document between separate find and
+    /// delete operations.
+    /// - Parameters:
+    ///   - filter: A `Document` that should match the query.
+    ///   - options: `FindOneAndModifyOptions` to use when executing the command.
+    /// - Returns: `Document` result of the attempt to delete a document  or `nil` if document wasn't found.
+    public func findOneAndDelete(filter: Document,
+                                 options: FindOneAndModifyOptions = .init(),
+                                 _isolation: isolated (any Actor)? = #isolation) async throws -> Document? {
+        try await withCheckedThrowingContinuation { continuation in
+            __findOneAndDeleteWhere(ObjectiveCSupport.convert(filter),
+                                    options: options) { bson, error in
+                if let error {
+                    continuation.resume(throwing: error)
+                } else {
+                    continuation.resume(returning: bson.map(ObjectiveCSupport.convert))
+                }
+            }
+        }
+    }
+#endif
 }
 
 private class ChangeEventDelegateProxy: RLMChangeEventDelegate {
