@@ -191,6 +191,7 @@ public protocol RealmSectionedResult: RandomAccessCollection, Equatable, ThreadC
                  _ block: @escaping (SectionedResultsChange<Self>) -> Void) -> NotificationToken
 }
 
+#if compiler(<6)
 public extension RealmSectionedResult {
     func observe(keyPaths: [String]? = nil,
                  on queue: DispatchQueue? = nil,
@@ -208,10 +209,34 @@ public extension RealmSectionedResult {
             collection.observe(keyPaths: keyPaths, on: nil) { change in
                 actor.invokeIsolated(block, change)
             }
-        } ?? NotificationToken()
+        }
     }
 }
+#else
+public extension RealmSectionedResult {
+    func observe(keyPaths: [String]? = nil,
+                 on queue: DispatchQueue? = nil,
+                 _ block: @escaping (SectionedResultsChange<Self>) -> Void) -> NotificationToken {
+        observe(keyPaths: keyPaths, on: queue, block)
+    }
 
+    /// :nodoc:
+    @available(macOS 10.15, tvOS 13.0, iOS 13.0, watchOS 6.0, *)
+    func observe<A: Actor>(
+        keyPaths: [String]? = nil, on actor: A,
+        _isolation: isolated (any Actor)? = #isolation,
+        _ block: @Sendable @escaping (isolated A, SectionedResultsChange<Self>) -> Void
+    ) async -> NotificationToken {
+        await with(self, on: actor) { actor, collection in
+            collection.observe(keyPaths: keyPaths, on: nil) { change in
+                actor.invokeIsolated(block, change)
+            }
+        }
+    }
+}
+#endif
+
+#if compiler(<6)
 public extension RealmSectionedResult where Element: RealmSectionedResult, Element.Element: ObjectBase {
     /**
      Registers a block to be called each time the sectioned results collection changes.
@@ -481,6 +506,279 @@ public extension RealmSectionedResult where Element: ObjectBase {
         await observe(keyPaths: keyPaths.map(_name(for:)), on: actor, block)
     }
 }
+#else
+public extension RealmSectionedResult where Element: RealmSectionedResult, Element.Element: ObjectBase {
+    /**
+     Registers a block to be called each time the sectioned results collection changes.
+
+     The block will be asynchronously called with the initial sectioned results collection, and then called again after each write
+     transaction which changes either any of the objects in the sectioned results collection, or which objects are in the sectioned results collection.
+
+     The `change` parameter that is passed to the block reports, in the form of indices within the collection, which of
+     the objects were added, removed, or modified during each write transaction. See the `SectionedResultsChange`
+     documentation for more information on the change information supplied and an example of how to use it to update a
+     `UITableView`.
+
+     At the time when the block is called, the collection will be fully evaluated and up-to-date, and as long as you do
+     not perform a write transaction on the same thread or explicitly call `realm.refresh()`, accessing it will never
+     perform blocking work.
+
+     If no queue is given, notifications are delivered via the standard run loop, and so can't be delivered while the
+     run loop is blocked by other activity. If a queue is given, notifications are delivered to that queue instead. When
+     notifications can't be delivered instantly, multiple notifications may be coalesced into a single notification.
+     This can include the notification with the initial sectioned results collection.
+
+     For example, the following code performs a write transaction immediately after adding the notification block, so
+     there is no opportunity for the initial notification to be delivered first. As a result, the initial notification
+     will reflect the state of the Realm after the write transaction.
+
+     ```swift
+     let dogs = realm.objects(Dog.self)
+     let sectionedResults = dogs.sectioned(by: \.age, ascending: true)
+     print("sectionedResults.count: \(sectionedResults?.count)") // => 0
+     let token = sectionedResults.observe { changes in
+         switch changes {
+         case .initial(let sectionedResults):
+             // Will print "sectionedResults.count: 1"
+             print("sectionedResults.count: \(sectionedResults.count)")
+             break
+         case .update:
+             // Will not be hit in this example
+             break
+         case .error:
+             break
+         }
+     }
+     try! realm.write {
+         let dog = Dog()
+         dog.name = "Rex"
+         person.dogs.append(dog)
+     }
+     // end of run loop execution context
+     ```
+
+     If no key paths are given, the block will be executed on any insertion,
+     modification, or deletion for all object properties and the properties of
+     any nested, linked objects. If a key path or key paths are provided,
+     then the block will be called for changes which occur only on the
+     provided key paths. For example, if:
+     ```swift
+     class Dog: Object {
+         @Persisted var name: String
+         @Persisted var age: Int
+         @Persisted var toys: List<Toy>
+     }
+     // ...
+     let dogs = realm.objects(Dog.self)
+     let sectionedResults = dogs.sectioned(by: \.age, ascending: true)
+     let token = sectionedResults.observe(keyPaths: ["name"]) { changes in
+         switch changes {
+         case .initial(let sectionedResults):
+            // ...
+         case .update:
+            // This case is hit:
+            // - after the token is initialized
+            // - when the name property of an object in the
+            // collection is modified
+            // - when an element is inserted or removed
+            //   from the collection.
+            // This block is not triggered:
+            // - when a value other than name is modified on
+            //   one of the elements.
+         case .error:
+             // ...
+         }
+     }
+     // end of run loop execution context
+     ```
+     - If the observed key path were `["toys.brand"]`, then any insertion or
+     deletion to the `toys` list on any of the collection's elements would trigger the block.
+     Changes to the `brand` value on any `Toy` that is linked to a `Dog` in this
+     collection will trigger the block. Changes to a value other than `brand` on any `Toy` that
+     is linked to a `Dog` in this collection would not trigger the block.
+     Any insertion or removal to the `Dog` type collection being observed
+     would also trigger a notification.
+     - If the above example observed the `["toys"]` key path, then any insertion,
+     deletion, or modification to the `toys` list for any element in the collection
+     would trigger the block.
+     Changes to any value on any `Toy` that is linked to a `Dog` in this collection
+     would *not* trigger the block.
+     Any insertion or removal to the `Dog` type collection being observed
+     would still trigger a notification.
+     - Any modification to the section key path property which results in the object changing
+     position in the section, or changing section entirely will trigger a notification.
+
+     - note: Multiple notification tokens on the same object which filter for
+     separate key paths *do not* filter exclusively. If one key path
+     change is satisfied for one notification token, then all notification
+     token blocks for that object will execute.
+
+     You must retain the returned token for as long as you want updates to be sent to the block. To stop receiving
+     updates, call `invalidate()` on the token.
+
+     - warning: This method cannot be called during a write transaction, or when the containing Realm is read-only.
+
+     - parameter keyPaths: Only properties contained in the key paths array will trigger
+                           the block when they are modified. If `nil`, notifications
+                           will be delivered for any property change on the object.
+     - parameter queue: The serial dispatch queue to receive notification on. If
+                        `nil`, notifications are delivered to the current thread.
+     - parameter block: The block to be called whenever a change occurs.
+     - returns: A token which must be held for as long as you want updates to be delivered.
+     */
+    func observe(keyPaths: [PartialKeyPath<Element.Element>],
+                 on queue: DispatchQueue? = nil,
+                 _ block: @escaping (SectionedResultsChange<Self>) -> Void) -> NotificationToken {
+        observe(keyPaths: keyPaths.map(_name(for:)), on: queue, block)
+    }
+
+    /// :nodoc:
+    @available(macOS 10.15, tvOS 13.0, iOS 13.0, watchOS 6.0, *)
+    func observe<A: Actor>(
+        keyPaths: [PartialKeyPath<Element.Element>], on actor: A,
+        _isolation: isolated (any Actor)? = #isolation,
+        _ block: @Sendable @escaping (isolated A, SectionedResultsChange<Self>) -> Void
+    ) async -> NotificationToken {
+        await observe(keyPaths: keyPaths.map(_name(for:)), on: actor, block)
+    }
+}
+
+public extension RealmSectionedResult where Element: ObjectBase {
+    /**
+     Registers a block to be called each time the sectioned results collection changes.
+
+     The block will be asynchronously called with the initial sectioned results collection, and then called again after each write
+     transaction which changes either any of the objects in the sectioned results collection, or which objects are in the sectioned results collection.
+
+     The `change` parameter that is passed to the block reports, in the form of indices within the collection, which of
+     the objects were added, removed, or modified during each write transaction. See the `SectionedResultsChange`
+     documentation for more information on the change information supplied and an example of how to use it to update a
+     `UITableView`.
+
+     At the time when the block is called, the collection will be fully evaluated and up-to-date, and as long as you do
+     not perform a write transaction on the same thread or explicitly call `realm.refresh()`, accessing it will never
+     perform blocking work.
+
+     If no queue is given, notifications are delivered via the standard run loop, and so can't be delivered while the
+     run loop is blocked by other activity. If a queue is given, notifications are delivered to that queue instead. When
+     notifications can't be delivered instantly, multiple notifications may be coalesced into a single notification.
+     This can include the notification with the initial sectioned results collection.
+
+     For example, the following code performs a write transaction immediately after adding the notification block, so
+     there is no opportunity for the initial notification to be delivered first. As a result, the initial notification
+     will reflect the state of the Realm after the write transaction.
+
+     ```swift
+     let dogs = realm.objects(Dog.self)
+     let sectionedResults = dogs.sectioned(by: \.age, ascending: true)
+     print("sectionedResults.count: \(sectionedResults?.count)") // => 0
+     let token = sectionedResults.observe { changes in
+         switch changes {
+         case .initial(let sectionedResults):
+             // Will print "sectionedResults.count: 1"
+             print("sectionedResults.count: \(sectionedResults.count)")
+             break
+         case .update:
+             // Will not be hit in this example
+             break
+         case .error:
+             break
+         }
+     }
+     try! realm.write {
+         let dog = Dog()
+         dog.name = "Rex"
+         person.dogs.append(dog)
+     }
+     // end of run loop execution context
+     ```
+
+     If no key paths are given, the block will be executed on any insertion,
+     modification, or deletion for all object properties and the properties of
+     any nested, linked objects. If a key path or key paths are provided,
+     then the block will be called for changes which occur only on the
+     provided key paths. For example, if:
+     ```swift
+     class Dog: Object {
+         @Persisted var name: String
+         @Persisted var age: Int
+         @Persisted var toys: List<Toy>
+     }
+     // ...
+     let dogs = realm.objects(Dog.self)
+     let sectionedResults = dogs.sectioned(by: \.age, ascending: true)
+     let token = sectionedResults.observe(keyPaths: ["name"]) { changes in
+         switch changes {
+         case .initial(let sectionedResults):
+            // ...
+         case .update:
+            // This case is hit:
+            // - after the token is initialized
+            // - when the name property of an object in the
+            // collection is modified
+            // - when an element is inserted or removed
+            //   from the collection.
+            // This block is not triggered:
+            // - when a value other than name is modified on
+            //   one of the elements.
+         case .error:
+             // ...
+         }
+     }
+     // end of run loop execution context
+     ```
+     - If the observed key path were `["toys.brand"]`, then any insertion or
+     deletion to the `toys` list on any of the collection's elements would trigger the block.
+     Changes to the `brand` value on any `Toy` that is linked to a `Dog` in this
+     collection will trigger the block. Changes to a value other than `brand` on any `Toy` that
+     is linked to a `Dog` in this collection would not trigger the block.
+     Any insertion or removal to the `Dog` type collection being observed
+     would also trigger a notification.
+     - If the above example observed the `["toys"]` key path, then any insertion,
+     deletion, or modification to the `toys` list for any element in the collection
+     would trigger the block.
+     Changes to any value on any `Toy` that is linked to a `Dog` in this collection
+     would *not* trigger the block.
+     Any insertion or removal to the `Dog` type collection being observed
+     would still trigger a notification.
+     - Any modification to the section key path property which results in the object changing
+     position in the section, or changing section entirely will trigger a notification.
+
+     - note: Multiple notification tokens on the same object which filter for
+     separate key paths *do not* filter exclusively. If one key path
+     change is satisfied for one notification token, then all notification
+     token blocks for that object will execute.
+
+     You must retain the returned token for as long as you want updates to be sent to the block. To stop receiving
+     updates, call `invalidate()` on the token.
+
+     - warning: This method cannot be called during a write transaction, or when the containing Realm is read-only.
+
+     - parameter keyPaths: Only properties contained in the key paths array will trigger
+                           the block when they are modified. If `nil`, notifications
+                           will be delivered for any property change on the object.
+     - parameter queue: The serial dispatch queue to receive notification on. If
+                        `nil`, notifications are delivered to the current thread.
+     - parameter block: The block to be called whenever a change occurs.
+     - returns: A token which must be held for as long as you want updates to be delivered.
+     */
+    func observe(keyPaths: [PartialKeyPath<Element>],
+                 on queue: DispatchQueue? = nil,
+                 _ block: @escaping (SectionedResultsChange<Self>) -> Void) -> NotificationToken {
+        observe(keyPaths: keyPaths.map(_name(for:)), on: queue, block)
+    }
+
+    /// :nodoc:
+    @available(macOS 10.15, tvOS 13.0, iOS 13.0, watchOS 6.0, *)
+    func observe<A: Actor>(
+        keyPaths: [PartialKeyPath<Element>], on actor: A,
+        _isolation: isolated (any Actor)? = #isolation,
+        _ block: @Sendable @escaping (isolated A, SectionedResultsChange<Self>) -> Void
+    ) async -> NotificationToken {
+        await observe(keyPaths: keyPaths.map(_name(for:)), on: actor, block)
+    }
+}
+#endif
 
 // Shared implementation of SectionedResults and ResultsSection
 private protocol SectionedResultImpl: RealmSectionedResult {
