@@ -19,15 +19,9 @@
 import XCTest
 import RealmSwift
 
-var pkCounter = 0
-func nextPrimaryKey() -> Int {
-    pkCounter += 1
-    return pkCounter
-}
-
 @available(*, deprecated) // Silence deprecation warnings for RealmOptional
 class SwiftKVOObject: Object {
-    @objc dynamic var pk = nextPrimaryKey() // primary key for equality
+    @objc dynamic var pk = ObjectId.generate() // primary key for equality
     @objc dynamic var ignored: Int = 0
 
     @objc dynamic var boolCol: Bool = false
@@ -194,13 +188,18 @@ class KVOTests: TestCase, @unchecked Sendable {
         changeDictionary = nil
     }
 
-    func observeChange<T: Equatable>(_ obj: SwiftKVOObject, _ keyPath: KeyPath<SwiftKVOObject, T>, _ old: T, _ new: T,
-                                     fileName: StaticString = #file, lineNumber: UInt = #line, _ block: () -> Void) {
+    func observeChange<T: Equatable>(
+        _ obj: SwiftKVOObject, _ keyPath: KeyPath<SwiftKVOObject, T>, _ old: T, _ new: T,
+        fileName: StaticString = #filePath, lineNumber: UInt = #line, _ block: () -> Void
+    ) {
+        // observe()'s callback is marked as Sendable, but we'll only ever call it from the same thread
         let kvoOptions: NSKeyValueObservingOptions = [.old, .new]
-        var gotNotification = false
+        nonisolated(unsafe) var gotNotification = false
+        nonisolated(unsafe) let nonisolatedOld = old
+        nonisolated(unsafe) let nonisolatedNew = new
         let observation = obj.observe(keyPath, options: kvoOptions) { _, change in
-            XCTAssertEqual(change.oldValue, old, file: (fileName), line: lineNumber)
-            XCTAssertEqual(change.newValue, new, file: (fileName), line: lineNumber)
+            XCTAssertEqual(change.oldValue, nonisolatedOld, file: fileName, line: lineNumber)
+            XCTAssertEqual(change.newValue, nonisolatedNew, file: fileName, line: lineNumber)
             gotNotification = true
         }
 
@@ -210,20 +209,24 @@ class KVOTests: TestCase, @unchecked Sendable {
         XCTAssertTrue(gotNotification, file: (fileName), line: lineNumber)
     }
 
-    func observeChange<T: Equatable>(_ obj: SwiftKVOObject, _ keyPath: KeyPath<SwiftKVOObject, T?>, _ old: T?, _ new: T?,
-                                     fileName: StaticString = #file, lineNumber: UInt = #line, _ block: () -> Void) {
+    func observeChange<T: Equatable>(
+        _ obj: SwiftKVOObject, _ keyPath: KeyPath<SwiftKVOObject, T?>, _ old: T?, _ new: T?,
+        fileName: StaticString = #filePath, lineNumber: UInt = #line, _ block: () -> Void
+    ) {
         let kvoOptions: NSKeyValueObservingOptions = [.old, .new]
-        var gotNotification = false
+        nonisolated(unsafe) var gotNotification = false
+        nonisolated(unsafe) let nonisolatedOld = old
+        nonisolated(unsafe) let nonisolatedNew = new
         let observation = obj.observe(keyPath, options: kvoOptions) { _, change in
             if let oldValue = change.oldValue {
-                XCTAssertEqual(oldValue, old, file: (fileName), line: lineNumber)
+                XCTAssertEqual(oldValue, nonisolatedOld, file: (fileName), line: lineNumber)
             } else {
-                XCTAssertNil(old, file: (fileName), line: lineNumber)
+                XCTAssertNil(nonisolatedOld, file: (fileName), line: lineNumber)
             }
             if let newValue = change.newValue {
-                XCTAssertEqual(newValue, new, file: (fileName), line: lineNumber)
+                XCTAssertEqual(newValue, nonisolatedNew, file: (fileName), line: lineNumber)
             } else {
-                XCTAssertNil(new, file: (fileName), line: lineNumber)
+                XCTAssertNil(nonisolatedNew, file: (fileName), line: lineNumber)
             }
             gotNotification = true
         }
